requirements.txt
flask
prometheus_client
requests
python-dotenv
